+++ packages/integration-test/importNamed.test.ts
@@ -19,6 +19,9 @@ import { iterableSome } from 'iter-fest/iterableSome';
 import { iterableToSpliced } from 'iter-fest/iterableToSpliced';
 import { iterableToString } from 'iter-fest/iterableToString';
 import { iteratorToIterable } from 'iter-fest/iteratorToIterable';
+import { Observable } from 'iter-fest/observable';
+import { observableValues } from 'iter-fest/observableValues';
+import { SymbolObservable } from 'iter-fest/symbolObservable';
 
 test('iterableAt should work', () => expect(iterableAt([1, 2, 3].values(), 1)).toBe(2));
 
@@ -88,3 +91,36 @@ test('iteratorToIterable should work', () =>
       )
     )
   ).toEqual([1, 2, 3]));
+
+test('Observable should work', () => {
+  const next = jest.fn();
+  const complete = jest.fn();
+
+  const observable = new Observable(observer => {
+    observer.next(1);
+    observer.complete();
+  });
+
+  observable.subscribe({ complete, next });
+
+  expect(next).toHaveBeenCalledTimes(1);
+  expect(next).toHaveBeenNthCalledWith(1, 1);
+  expect(complete).toHaveBeenCalledTimes(1);
+});
+
+test('observableValues should work', async () => {
+  const observable = Observable.from([1, 2, 3]);
+  const values = [];
+
+  for await (const value of observableValues(observable)) {
+    values.push(value);
+  }
+
+  expect(values).toEqual([1, 2, 3]);
+});
+
+test('SymbolObservable should work', () => {
+  const observable = new Observable(() => {});
+
+  expect(observable[SymbolObservable]()).toBe(observable);
+});